refactor: colocate fallback handoff with client lifecycle
